+++ packages/client/src/Call.ts
@@ -152,14 +152,14 @@ export class Call {
   watching: boolean;
 
   /**
-   * The permissions context of this call.
-   */
-  readonly permissionsContext = new PermissionsContext();
-
-  /**
    * Flag telling whether this call is a "ringing" call.
    */
   private readonly ringingSubject: Subject<boolean>;
+
+  /**
+   * The permissions context of this call.
+   */
+  readonly permissionsContext = new PermissionsContext();
 
   /**
    * The event dispatcher instance dedicated to this Call instance.