+++ packages/react-bindings/src/contexts/StreamVideoContext.tsx
@@ -1,4 +1,10 @@
-import { createContext, PropsWithChildren, useContext } from 'react';
+import {
+  createContext,
+  PropsWithChildren,
+  useContext,
+  useEffect,
+  useRef,
+} from 'react';
 import { StreamVideoClient } from '@stream-io/video-client';
 import {
   StreamI18nProvider,
@@ -30,6 +36,26 @@ export const StreamVideo = ({
   language,
   translationsOverrides,
 }: PropsWithChildren<StreamVideoProps>) => {
+  const prevClient = useRef<StreamVideoClient | undefined>(undefined);
+
+  useEffect(() => {
+    if (client.user) {
+      client
+        .connectUser()
+        .catch((error) =>
+          console.error('Failed to establish connection', error),
+        );
+    }
+
+    prevClient.current = client;
+
+    return () => {
+      prevClient.current
+        ?.disconnectUser()
+        .catch((error) => console.error(`Failed to disconnect`, error));
+    };
+  }, [client]);
+
   return (
     <StreamVideoContext.Provider value={client}>
       <StreamI18nProvider