+++ packages/react-native-dogfood/src/screens/Meeting/GuestMeetingScreen.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import {
   StreamCall,
   StreamVideo,
-  TokenOrProvider,
   User,
   useCreateStreamVideoClient,
 } from '@stream-io/video-react-native-sdk';
@@ -19,11 +18,10 @@ type Props = NativeStackScreenProps<
 export const GuestMeetingScreen = (props: Props) => {
   const apiKey = process.env.STREAM_API_KEY as string;
   const {
-    params: { guestUserId, guestCallId, mode },
+    params: { guestUserId, callId, mode },
   } = props.route;
-  const guestCallType = 'default';
+  const callType = 'default';
 
-  const [tokenToUse, setTokenToUse] = useState<TokenOrProvider>(undefined);
   const [show, setShow] = useState<ScreenTypes>('lobby');
   const { navigation } = props;
 
@@ -41,6 +39,21 @@ export const GuestMeetingScreen = (props: Props) => {
           },
     [mode, guestUserId],
   );
+
+  const tokenOrProvider = useCallback(async () => {
+    const token = await createToken({
+      user_id: '!anon',
+      call_cids: `${callType}:${callId}`,
+    });
+    return token;
+  }, [callId, callType]);
+
+  const client = useCreateStreamVideoClient({
+    apiKey,
+    tokenOrProvider: mode === 'guest' ? undefined : tokenOrProvider,
+    user: userToConnect,
+  });
+
 
   const onJoin = () => {
     setShow('active-call');
@@ -72,19 +85,14 @@ export const GuestMeetingScreen = (props: Props) => {
   return (
     <StreamVideo client={client}>
       <StreamCall
-        callId={guestCallId}
-        callType={guestCallType}
+        callId={callId}
+        callType={callType}
         callCycleHandlers={{
           onCallJoined: onJoin,
           onCallHungUp: onLeave,
         }}
       >
-        <MeetingUI
-          show={show}
-          setShow={setShow}
-          callId={guestCallId}
-          {...props}
-        />
+        <MeetingUI show={show} setShow={setShow} callId={callId} {...props} />
       </StreamCall>
     </StreamVideo>
   );