+++ sample-apps/react/react-video-demo/tailwind.config.cjs
@@ -14,10 +14,7 @@ const baseFallbackFonts = [
   'sans-serif',
 ];
 module.exports = {
-  content: [
-    "./src/components/**/*.{html, js, ts, tsx}",
-    "./src/**/*"
-  ],
+  content: ['./src/components/**/*.{html, js, ts, tsx}', './src/**/*'],
   theme: {
     borderWidth: {
       default: '1px',
@@ -61,9 +58,9 @@ module.exports = {
           blue: '#2F7DEB',
         },
         popup: {
-          description: '#72767E'
-        }
-      }
+          description: '#72767E',
+        },
+      },
     },
     gradients: {
       button: ['180deg', 'rgba(0, 138, 255, 0.95)', 'rgba(0, 138, 255, 1)'],
@@ -74,29 +71,31 @@ module.exports = {
       DEFAULT: '0 2px 18px 0 rgba(0,0,0,0.04)',
       strong: '0 2px 18px 0 rgba(0,0,0,0.1)',
       blue: '0 2px 19px rgba(0, 0, 0, 0.25), 0 7px 32px #21B5E3, inset 0 1px 23px 0 rgba(8, 231, 255, 0.2)',
-      button: '0 1px 1px 0 rgba(0, 94, 173, 1), 0 3px 13px 0 rgba(0, 125, 238, 0.32)',
+      button:
+        '0 1px 1px 0 rgba(0, 94, 173, 1), 0 3px 13px 0 rgba(0, 125, 238, 0.32)',
       'button-default': '0 3px 13px 0 rgba(0,0,0,0.32)',
       'button-hover': '0 3px 13px 1px rgba(0,0,0,0.35)',
       'button-light': '0 3px 13px 0 rgba(0,0,0,0.08)',
       icon: '0 2px 4px 0 rgba(0,0,0,0.06), 0 7px 32px 0 rgba(33,181,227,0.28)',
-      popup: '0 50px 100px rgba(50, 50, 93, 0.1), 0 15px 35px rgba(50, 50, 93, 0.15), 0 5px 15px rgba(0, 0, 0, 0.1)',
+      popup:
+        '0 50px 100px rgba(50, 50, 93, 0.1), 0 15px 35px rgba(50, 50, 93, 0.15), 0 5px 15px rgba(0, 0, 0, 0.1)',
       sidebar: '0 2px 30px 0 rgba(0,0,0,0.10)',
       'product-card': '0 2px 30px rgba(0,0,0,0.1)',
       'demo-card': '0 0 25px rgba(0,0,0,0.26)',
-      footer: 'inset 0 1px 0 0 rgba(0,0,0,0.07), inset 0 8px 7px 0 rgba(0,0,0,0.01)',
+      footer:
+        'inset 0 1px 0 0 rgba(0,0,0,0.07), inset 0 8px 7px 0 rgba(0,0,0,0.01)',
       dialog: '0 5px 10px 0 rgba(0,0,0,0.06)',
       main: '0px 0px 30px 12px rgb(238 238 238)',
       card: '0px 0px 3px 0px rgba(0,0,0,0.06)',
       cardActive: '0px 0px 20px 0px rgba(0,0,0,0.06)',
     },
 
-
     extend: {
       borderRadius: {
         panel: '13px', // rounded-container // rounded-demo-container
         button: '8px',
         container: '13px',
-        input: '6px'
+        input: '6px',
       },
       colors: {
         video: {
@@ -120,7 +119,7 @@ module.exports = {
             500: '#1C1E22',
             600: '#212326',
             700: '#272A30',
-            800: '#121416'
+            800: '#121416',
           },
           green: {
             500: '#20E070',
@@ -134,11 +133,12 @@ module.exports = {
           yellow: {
             500: '#FFB700',
           },
-        }
-      }
+        },
+      },
     },
   },
-  plugins: [tailwindGradients,
+  plugins: [
+    tailwindGradients,
     //   function ({
     //   addBase,
     //   theme